# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ms-CRM OpenApi Documentation",
        "description": "ms-CRM OpenApi description",
        "contact": {
            "email": "florian@alt.bzh"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/",
            "description": "Full ms-CRM API endPoint documentation"
        }
    ],
    "paths": {
...
        "/money/company/{id}": {
            "get": {
                "tags": [
                    "VirtualMoney"
                ],
                "summary": "get sold for Virtual Money of this company",
                "description": "Affichage du solde du compte virtuel d'une société",
                "operationId": "67228d6bf83501096bb0397b2ebcc155",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return VirtualMoney object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VirtualMoney"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No account found"
                    }
                }
            }
        },
        "/money/user/{id}": {
            "get": {
                "tags": [
                    "VirtualMoney"
                ],
                "summary": "get sold for Virtual Money of this user",
                "description": "Affichage du solde du compte virtuel d'un user",
                "operationId": "e65ff7e7ed137641f69264283fd78d6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return VirtualMoney object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VirtualMoney"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No account found"
                    }
                }
            }
        },
        "/money/company": {
            "post": {
                "tags": [
                    "VirtualMoney"
                ],
                "summary": "Store a Virtual Money Account",
                "description": "Création d'un compte virtuel pour une société",
                "operationId": "b1b14799e10b3b869f740e6ae50e4c80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return VirtualMoney object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VirtualMoney"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not store"
                    }
                }
            }
        },
        "/money/user": {
            "post": {
                "tags": [
                    "VirtualMoney"
                ],
                "summary": "Store a Virtual Money Account",
                "description": "Création d'un compte virtuel pour un user",
                "operationId": "e100bd94e7c55334abba15686dc21085",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return VirtualMoney object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VirtualMoney"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not store"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CountRequest": {
                "title": "CountRequest",
                "description": "Store Account Virtual Money request body data",
                "required": [
                    "type",
                    "type_id"
                ],
                "properties": {
                    "type": {
                        "title": "type",
                        "description": "1 = company, 2 = user",
                        "type": "integer",
                        "example": "2"
                    },
                    "reference_id": {
                        "title": "reference_id",
                        "description": "id of user or company",
                        "type": "integer",
                        "example": "2"
                    }
                },
                "type": "object"
            },
...
            "VirtualMoney": {
                "title": "VirtualMoney",
                "description": "VirtualMoney model",
                "properties": {
                    "type": {
                        "title": "type",
                        "description": "1 = company, 2 = user",
                        "type": "integer",
                        "example": "2"
                    },
                    "amount": {
                        "title": "amount",
                        "description": "amount of this exchange",
                        "type": "integer",
                        "example": "2"
                    },
                    "reference_id": {
                        "title": "reference_id",
                        "description": "id of user or company",
                        "type": "integer",
                        "example": "2"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "VirtualMoney"
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}